tests/amdgpu/amd_vrr_range: Skip on unsupported connector types
edid_database includes a set of EDID but does not includes one with
connector type == DRM_MODE_CONNECTOR_WRITEBACK. With writeback connector
enabled, this will always fail even if writeback connector does not
support freesync. As a result, freesync tests is skipped when there is
no match in edid_database.

Signed-off-by: Alex Hung <alex.hung@amd.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>