tests/intel/kms_dsc: Use high resolution mode for output format subtests
Currently, we are not using the highest resolution mode for output formats
other than RGB. For testing other formats we force an output format, and
try the mode from list of modes.
Lets call get_highres_mode before using the modes, so that we try modes
that are sorted in descending order of resolution. This is to ensure we
try with higher resolution mode and use DSC with that mode.

Signed-off-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Reviewed-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>