tests/intel: Use MI_ARB_CHECK definition more consistently
Even though we have MI_ARB_CHECK defined in our GPU instruction header,
several of our tests and test libraries seem to be using alternate
definitions and/or magic numbers to emit this instruction, which makes
grep'ing the code for it challenging (and makes the behavior of the
tests themselves less obvious in some cases).  Try to use the standard
definition more consistently everywhere.

Reviewed-by: Gustavo Sousa <gustavo.sousa@intel.com>
Link: https://lore.kernel.org/r/20240328174139.1533658-1-matthew.d.roper@intel.com
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>