tools/gputop: Fix engine columns with amdgpu
Amdgpu kernel driver skips output of fdinfo keys for unused engines. That
is completely legal but corrupts the column formatting in the current
code.

Fix it by simply treating change in detected engines used by a client
as trigger to re-emit a new header. This ensures columns are always
correctly aligned, albeit with a cost of potentially duplicating the
header for the same DRM minor.

This is considered good enough for a reference implementation. The
alternative would be to add some real per DRM minor state tracking which
sounds like an overkill, at least until gputop gains a nicer (any) UI.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@igalia.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Acked-by: Christian König <christian.koenig@amd.com>