tests/intel/xe_exec_store: fix sync usage
If using async binds it looks like an in-fence for the exec is needed to
ensure the exec happens after the out-fence from the binds are complete.
Therefore we need to unset DRM_XE_SYNC_FLAG_SIGNAL after doing the
binds, but before the exec, otherwise the sync is rather treated
as an out-fence and the binds can then happen after the exec, leading to
various failures. In addition it looks like async unbind should be
waited on before tearing down the queue/vm which has the bind engine
attached, since the scheduler timeout is immediately set to zero on
destroy, which might then trigger job timeouts. However it looks like
it's also fine to rather just destroy the object and leave KMD to unbind
everything itself. Update the various subtests here to conform to this.

In the case of the persistent subtest it looks simpler to use sync
vm_bind since we don't have another sync for the in-fence at hand, plus
we don't seem to need a dedicated bind engine.

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/1270
Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>