tests/intel/xe_exec_fault_mode: Fix misuse of __xe_wait_ufence
When confirming that xe_exec() user_fences completed, this is performed
in a loop.  When using __xe_wait_ufence(), it returns an updated timeout
value to reflect how much time was remaining.  As side-effect, when calling
in a loop without resetting the timeout, the timeout value will get
progressively smaller over many iterations.
We don't want a smaller timeout to be used on each subsequent call to
__xe_wait_ufence(), so fix here is to reset the timeout inside the loop.
This issue is visible in simulation runs due to the slower execution time,
where later iterations of the wait are failing due to timeout being too
small.

Signed-off-by: Brian Welty <brian.welty@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>