tools/intel_vbt_decode: Use .min_bdb_version to filter out PSR block on pre-165 VBTs
Block 9 on ALM was something different, then disappeared until
it got repurposed to serve as the PSR block starting from
version 165.

Utilize the new .min_bdb_version to filter out this block on
older VBTs. We'll make more use of this mechanism when we start
decoding the other side of this coin.

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>