tools/intel_vbt_decode: Decode the "not HDMI" bit right way up
The inverted decoding of the "not HDMI" bit is confusing me more
than it's helping as it typically gets flagged for all non-HDMI/DVI
encoders now. Just decode it the same way the VBT specifies
it.

The other option would be to only do the invert trick when TMDS
encoding is supported. But dunno how confusing that would be.

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>