tools/intel_vbt_decode: Use "(LFP<n>)" to indicate LFP panel type
The normal panel type is for "LFP" panels. It is also possible
that some of the LFP blocks (IIRC blocks 43 and 44) would
also apply to EFP panels (declared via the child devices' efp_index).
To make that somewhat possible in the future use the "(LFP<n>)" format
when indicating the LFP panel type match.

TODO: figure out how to dig out the index/indices and indicate
which LFP data applies to which EFP panels...

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>