lib/igt_drm_clients: lazy stat process
Only try to get task data like pid and name when that is required.
Typically most of the processes in a system don't have a drm fd open.
So only open and use /proc/<pid>/stat when it's needed.

For the data below, system is running with just a single drm client:

Before:
 Performance counter stats for './build/tools/gputop -d 0.1 -n 10':

            507.19 msec task-clock                       #    0.334 CPUs utilized
                29      context-switches                 #   57.178 /sec
                 4      cpu-migrations                   #    7.887 /sec
                75      page-faults                      #  147.874 /sec
       (...)
       1.518253637 seconds time elapsed
       0.007989000 seconds user
       0.503351000 seconds sys

After:
 Performance counter stats for './build/tools/gputop -d 0.1 -n 10':

            384.01 msec task-clock                       #    0.275 CPUs utilized
                17      context-switches                 #   44.270 /sec
                 1      cpu-migrations                   #    2.604 /sec
                74      page-faults                      #  192.705 /sec
       (...)
       1.397640800 seconds time elapsed
       0.008106000 seconds user
       0.385048000 seconds sys

Looking only at the execution time of igt_drm_clients_scan() over 10
runs.
Before: 47138890 nsec
After:  33978388 nsec

So in both methods it shows a ~25% speedup. It is expected that in a
system with more clients this won't be as high.

Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@igalia.com>
Link: https://lore.kernel.org/r/20240501174001.121324-1-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>