tests/intel/xe_pm: S4 to go up to devices only
Testing S4 (hibernation) is typically painful as there's mixed support
in OS versions and platforms in CI. Doing the entire dance of saving the
image to swap (which sometimes is a swapfile) and communicate that to
the kernel that is going to be booted (without initrd in the CI case) is
often a case of problems.

Main goal of xe_pm is to test if the xe driver and the graphics card are
working correctly, not that all the farm of machines correctly handle
all the corner cases (which is even more problematic as we test early
rc kernels).

Stop doing that and rather switch to going up to device shutdown +
platform low power state (the default in /sys/power/disk). If that is
acceptable and work out great, we may even do that unconditionally,
passing SUSPEND_TEST_DEVICES as it should work in other cases too.

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/1043
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Link: https://lore.kernel.org/r/20240503224745.14890-3-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>