tests/kms_ccs: Test XRGB2101010
I want to enable 10bpc CCS support for i915. To that end start
testing it.

Note that we are faking things a bit because rendercopy will
still use a 8:8:8:8 format here (plumbing the full pixel format,
as opposed to just bpp, would require quite a bit more work).
However this seems to work fine, I suspect due to:
- we only do 1:1: copies of data
- the compression is lossless

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>