lib/igt_kms: move setting DRM_CLIENT_CAP_WRITEBACK_CONNECTORS to kms_writeback
Currently DRM_CLIENT_CAP_WRITEBACK_CONNECTORS is set in the igt_display_require()
which is invoked for all IGT tests irrespective of whether the rest of the
writeback connector properties such as the FB_ID are set.

For the writeback connectors to function properly, additional setup steps are
required (like setting up the output buffers and submitting the job). These steps
are not a part of the default IGT setup, so there is no guarantee that the pipeline
will be executed at all. This results in regressions for msm boards across multiple
IGT tests on writeback connectors such as CTM related tests where CRC needs to
be collected after the CTM has been applied in hardware.

It is unclear whether this is intentional to be able to run all IGT tests even
with writeback connector or it is only kms_writeback which is supposed to
be the one.

Lets try with the latter approach by setting DRM_CLIENT_CAP_WRITEBACK_CONNECTORS
only within kms_writeback.

changes since RFC:
	- minor fixes to commit message

changes in v2:
	- use igt_require_f to print descriptive message
	- add msm card failure details in commit text

Reviewed-by: Rob Clark <robdclark@gmail.com>
Acked-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Acked-by: Petri Latvala <adrinael@adrinael.net>
Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>