lib/xe/xe_query: return errno from xe_supports_faults check
Return the exact error value from the ioctl used in xe_supports_faults.
This is done because it is possible that faults are supported on the
system but are temporarily unavailable due to VM usage.  This change
allows the user to determine if they are willing to wait for fault
support to become active again.

Signed-off-by: Jonathan Cavitt <jonathan.cavitt@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>