lib/igt_drm_fdinfo: Parse drm-total-cycles
Parse drm-total-cycles and annotate what are the possible engine
utilization types that can be used. Currently there are 2 possible ways
to calculate utilization:

1) via drm-engine, that uses the time and derive the utilization by
   comparing to the CPU time
2) via drm-cycles + drm-total-cycles, that uses the GPU time

igt_drm_fdinfo just marks the presence of the keys, letting the caller
decide which method to use.

It is assumed that if any drm-engine, drm-cycles or drm-total-cycles is
available, it's available for all engines with non-zero utilization.

Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Link: https://lore.kernel.org/r/20240504064643.25863-12-lucas.demarchi@intel.com
Link: https://lore.kernel.org/r/20240521161525.159720-4-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>