tests/intel/xe_create: Don't assert on presi execution
create_execqueues subtest is testing the time required for creating
execqueues along with testing the functionality.
Presi environments are not for testing performance. We can't use the
same time limit set for silicon on presi environments as presi would be
slow.
Therefore the time limit check is skipped while executing on simulation.

Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>